requirements.txt
click==6.7
Flask==0.12
future==0.16.0
itsdangerous==0.24
Jinja2==2.9.4
line-bot-sdk==1.0.2
MarkupSafe==0.23
python-dotenv==0.6.2
requests==2.12.4
Werkzeug==0.11.15
